infraestructura_az/AuthWS/requirements.txt
tornado==4.3
pyjwt==2.1.0
python-ldap==3.3.1